+++ src/commonMain/kotlin/cengine/lang/asm/elf/elf64/ELF64_Shdr.kt
@@ -9,7 +9,7 @@ data class ELF64_Shdr(
     var sh_addr: Elf64_Addr = 0U,
     var sh_offset: Elf64_Off = 0U,
     var sh_size: Elf_Xword = 0U,
-    override var sh_link: Elf_Word = Shdr.SHN_UNDEF,
+    override var sh_link: Elf_Word = Shdr.SHN_UNDEF.toUInt(),
     override var sh_info: Elf_Word = 0U,
     var sh_addralign: Elf_Xword = 0U,
     var sh_entsize: Elf_Xword = 0U
